Improved readers and loaders leniency
The loaders and readers determine if the model/source provided is a three
dimensional or two dimensional view. Once the decision is made, the
reader/loader will convert the source into a model/source. This behavior
is normally found in modelLoader and ModelFormat because OpenModel is a
wrapper for the objects.